refactor: extract shared HTTP/SSE client to agent-core module
Both modem-agent and smpp-agent had near-identical implementations of
the Vendel backend client (FetchPending, ReportStatus, ReportIncoming,
ConnectSSE). Extract the common code into a new agent-core/ module
consumed by both agents via a local replace directive.

ConnectSSE now takes a topicPrefix parameter ("modem" or "smpp")
instead of hardcoding the topic. Dockerfiles and docker-compose build
contexts move up to the repo root so agent-core/ is reachable during
image builds.
