+++ agent-core/go.mod
@@ -1,0 +1,3 @@
+module github.com/JimScope/vendel/agent-core
+
+go 1.26.1
