+++ docker-compose.yml
@@ -30,7 +30,9 @@ services:
       - dev
 
   smpp-agent:
-    build: ./smpp-agent
+    build:
+      context: .
+      dockerfile: smpp-agent/Dockerfile
     environment:
       - VENDEL_URL=${VENDEL_URL:-http://app:8090}
       - CONFIG_FILE=/etc/smpp-agent/configs.yaml
